❇️[feat] 알림 SSE 응답 필드 추가 #200
