+++ src/main/java/com/daon/backend/notification/dto/FindNotificationsResponse.java
@@ -1,7 +1,5 @@
 package com.daon.backend.notification.dto;
 
-import com.daon.backend.notification.domain.Notification;
-import com.daon.backend.notification.domain.NotificationType;
 import lombok.Getter;
 
 import java.util.List;
@@ -13,17 +11,5 @@ public class FindNotificationsResponse {
 
     public FindNotificationsResponse(List<NotificationSummary> notifications) {
         this.notifications = notifications;
-    }
-
-    @Getter
-    public static class NotificationSummary {
-        private String data;
-
-        private NotificationType type;
-
-        public NotificationSummary(Notification notification) {
-            this.data = notification.getNotificationData();
-            this.type = notification.getNotificationType();
-        }
     }
 }
